Rename file to fix broken next config
